Rework imgcache module
